+++ blog-app/src/components/Layout.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Header from "../Home/Header";
+import Header from "./Header";
 import { Outlet } from "react-router-dom";
 
 function Layout() {
