+++ blog-app/src/styles/PostPage.css
